fix(blockchain-api): review follow-ups for batch expiry, migration race, e2e helpers
- batch-status: a confirmed-but-not-finalized transaction is in a block the
  cluster will not drop, so its blockhash ageing out no longer writes it
  expired (and its batch failed) when polled at finalized
- migration: serialise the concurrent index build across replicas with a
  session advisory lock on one pooled client; drop an invalid leftover
  CONCURRENTLY instead of taking ACCESS EXCLUSIVE on pending_transactions
- e2e workflow: run on edits to the workflow itself so the matrix-coverage
  check fires on a deleted matrix entry
- extend: judge lockup, expiration and season on the registrar clock
  (cluster clock plus time_offset), matching delegate
- e2e helpers: one checked setSurfnetAccount replaces seven surfnet_setAccount
  fetches, so a failed write (e.g. the time-offset reset) throws
- getMultipleAccounts chunking helper moves to lib/utils
